Changed linked list data structure
The data structure now consists of nodes that store the index of the
next and previous node (as opposed to refrences to the vertex & bisector
of the next and previous node).

Edge events are working ish, Split events are commented out.
The time variable of events are not stored anywhere at the moment.