Bug fix for handle_vertex_event() method: implemented a more strict criterium for identifying the edge beeing split.
while serching for the edge beeing split check if the
split point b lies inside the area bounded by the bisectors of the two
vertices composing the edge.

This criteia fixes a bug where the handle_split_event method would
incorectly identify a edge if b happens to lie on the line as the edge,
but outside the segment bounded by the end vertices of the edge.