Bug Fix: Incorect bisector calculation for vertices created form split events.
The current location of nodes used in the bisector function were
calculated incorrectly and as a result the bisector for newly created
nodes form split events were incorrect. The location of the points did
not consider the time of creation of the point in the current location
calculation.

The algorithm now seems to be working without error at least for all
contours extracted from the bunny model.