Bug Fix in Straight skeleton: compute_split_events incorrectly disguards split events for short edges.
In the compute_split_event function edge_vec
would sometimes be inverted for small polygons.
This results in incorrect calculation of the
split point: b. b would usually lie outside the valid
split zone resulting in the event being discarded.

In order to allow for multiple nodes to share a common
vertex edge_vec was calculated by offsetting each edge
vertex by its bisector. This trick makes it possible to
create a valid edge_vec between nodes that share a
common vertex. However this way of calculating edge_vec
would sometimes result in the vector beeing inverted for
short edges.

The this patch only uses the coordinates of the vertices
of the edge to construct edge_vec, and falls back to the
previous method (offsetting by bisectors) if the nodes of
the edge share the same vertex.