Bug Fix: straight skeleton compute_split_event
Previously the compute_split_event method would identify split events
allong the reflex vertex's negative bisector (as if the reflex vertex
was traveling in the opposite direction).

This patch adds a new check to see whether the split point b lies in
the same direction as the reflex vertex's bisector, to filter out
invalid split events.

The intersect function was also reworked in this patch to return a
optional intersection. This is aposed to the previous implementation
where the function would panic if no intersection could be found (due to
parallel vectors in the input). This change became nessesary as filtering
out the invalid split events lead to the intersect function panicing
often due to parallel vectors in the input (Not entierly sure why this is
the case). The new intersect funtion is likely also more effisient as it
does not use matrix inverse.