Messing with genereating support profiles
The support profiles are the 2D contours that define
where planar layers are adaquite. The remaining area inside
the mesh and not enclosed by a support profile are areas where
non-planar layers are needed.

The prosess works about 20% of the time. The remaining 80% of the time
no support profiles are generated or only half of the profiles are
generated. This is the case even when running the program on the same
input mesh.