extract_planar_layers improvement/bugfix + minor code cleanup
extract_planar_layers now does not generate a layer at z-height = 0
as this would create a layer that intersects with the build plate of
the printer.

extract_layer: replaced the old method for removing redundant vertices.
Previously a vertex would not get added to the contour if its proximity
to the previous point was less than a threshold. The new method uses the
area of the triangle formed by 3 neighbouring vertices as a criterium to
determine if a vertex should be included in the contour or not. The new
method in contrast to the old proximity method is able to remove redundant
vertices on parrallel lines.

edge_zplane_intersection: previously yealded a intersection point if the
z-plane intersected either of the two edge vertices. This caused problems
when attempting to extract_planar_layers when the input mesh contains
faces prallel to and at the same height as the z-plane. In this senario
the previous method would yeald overlapping or broken contours. This
problem was fixed by making the edge_zplane_intersection function yeald
no intersection when the z-plane intersects the edge vertex with the
lowest z-coorinate (while an intersection is still yealded for
intersections with the edge vertex with the highest z-cooridnate).