Initial implementation of bounded skeleton:
This patch adds support for computing a straight skeleton
inside a bounding contour. The current implementation
achives this by introdusing a bound event which is defined
by a node traveling outside the bounding contour. The bound
event is handled by simply deactivating the node and adding
the nesseary edges and vertices to the skeleton output.
This expreimental approach might be too simple and will
likely require rework in the future.