Added method for computing a straight skeleton up untill a time/distance limit.
The new compute_skeleton_with_limit mehtod computes
the straight skeleton untill the spesified time/distance
limit. The active nodes of the skeleton at the time limit
are added to the generated skeleton this makes it possible
to generate a mesh from the resulting StraightSkeleton.

Also implemented StraightSkeleton::from(SkeletonBuilder).