Initial implementation VertexTree and SkeletonBuilder::offset_inside_bounds
VertexTree is a tree structure that maps each trajectory/
propagation of vertices in a straight skeleton's input
contours to the offset contour of the straight skeleton.

SkeletonBuilder::offset_inside_bounds is a method to
generate a partial polygon offset from a SkeletonBuilder.
The active vertices of the SeletonBuilder that are inside
a bounding polygon are offset while the remaining vertices
keep their original location.