Bug Fix: Fixed off by one error extract_planar_layers() contour-simplification algorithm.
The bug was located in stl_op::extract_layer in the part of
the function that simplifies the extracted contour by removing
redundant verices. The bug caused the simplification algorithm
to calculate the incorect area of the triangle spaned by 3
neighbouring vertices. As this area is used as the deletion
criterium this bug caused the wrong vertices to be deleted.