Refactor Polygon struct to be a tupple struct of Vec<Contor>
This makes it easyer to loop over all contours in the stuct
which is constantly throughout the code. This also makes the
stuct more compattibe with i_overlay's Shape type.

The new Polygon struct implements Deref to Vec<Contour> this
might have been a misstake, and could be removed in the future.