Added Straigth-Skeleton-offset (ss_offset) method for polygons.
ss_offset uses the Straight Skelton implementation
in this project to performe polygon offseting. This
is an alternative to the other offset function witch
uses i_overlay for the implementation.
the new method was nessesary to avoid a problem with
the i_overlay method where additional points are sampled
on the output curve (undesiered behaviour).