Added functionality for rust to communicate to blender over TCP.
This patch includes functionality for createing a
TCP connection to blender. This is acheved by
launching blender with a python script (server.py).
The python script listens on the TCP connection and
parses mesh data encoded in json. This data is then
used to recreate mesh objects in blender. This
workflow is a replacement from the previous janky
method of debugging in utils.rs, where rust transfers
mesh data by creating a python library witch is read
by blender for mesh construction.