Move from cargo2nix to native nix devshell with oxalica rust overlay to get newer rustc version
A more updated rust compiler was needed in order
to use a newer version of the i_overlay crate.
This change is in preparation for upcomming
tagged_boolean functionality which requires features
from i_overlay version 3.0 to work.

The new flake.nix file does not provide a build output
as it did previously, as a result nix run command will
no longer work.